Implemented PeerStatus Machine For Peer Handshakes And Connections
