+++ bip_handshake/Cargo.toml
@@ -15,10 +15,9 @@ license       = "MIT/Apache-2.0"
 
 [dependencies]
 bip_util      = { path = "../bip_util", version = "0.4.0" }
-mio           = "0.5.0"
-slab          = "0.2.0"
 nom           = "1.2.0"
-log           = "0.3.0"
+rotor         = "0.6.0"
+rotor-stream  = "0.6.0"
 
 [features]
 unstable      = []
