+++ bip_handshake/src/lib.rs
@@ -1,15 +1,13 @@
 extern crate bip_util;
 #[macro_use]
-extern crate log;
-extern crate mio;
-#[macro_use]
 extern crate nom;
-extern crate slab;
+extern crate rotor;
+extern crate rotor_stream;
 
 mod bittorrent;
 mod handshaker;
 
 pub use handshaker::Handshaker;
-pub use bittorrent::{BTHandshaker, BTPeer};
+//pub use bittorrent::{BTHandshaker, BTPeer};
 
 pub use bip_util::bt::{PeerId, InfoHash};
